function cancelVote controller

+++ api/controller/StatsManager.php
@@ -100,6 +100,7 @@ class StatsManager {
 
   public function upVote($id, $name) // $id de la publication, $name de la stat
   {
+    $name--;
     if ($this->hasVoted($id, $name)) {
       return false; // l'user a déjà voté sur cette stat
     }
@@ -111,24 +112,20 @@ class StatsManager {
 
       // récupérer l'auteur de la publication
       $q2 = $this->_db->query('SELECT user_id FROM publication WHERE id = "'.$id.'"');
-      $donnees[] = $q2->fetch(PDO::FETCH_ASSOC);
-      var_dump($donnees);
+      $donnees = $q2->fetch(PDO::FETCH_ASSOC);
       $user_manager = new UsersManager($this->_db);
-      var_dump($user_manager);
       $user = $user_manager->get($donnees['user_id']);
-      var_dump($user);
       
       // update les stats de l'auteur
       $q3 = $this->_db->prepare('UPDATE stat SET value = value+1
-        WHERE related_user_id = "'.$user.'"
+        WHERE related_user_id = "'.$user->get_id().'"
         AND name = "'.$name.'"');
 
       // change la table vote
-      $q4 = $this->_db->prepare('INSERT INTO vote (user_id, publication_id, name)
-        VALUES ("'.$user.'", "'.$id.'", "'.$name.'"');
+      $q4 = $this->_db->prepare('INSERT INTO vote(user_id, publication_id, name)
+        VALUES ("'.$user->get_id().'", "'.$id.'", "'.$name.'")');
 
       $q1->execute();
-      $q2->execute();
       $q3->execute();
       $q4->execute();
       return true;
@@ -137,30 +134,36 @@ class StatsManager {
 
   public function cancelVote($id, $name)
   {
-    $q1 = $this->_db->prepare('UPDATE stat SET value = value-1
-      WHERE related_publication_id = "'.$id.'"
-      AND name = "'.$name.'"');
+    if ($this->hasVoted($id, $name)) {
+      $name--;
+      $q1 = $this->_db->prepare('UPDATE stat SET value = value-1
+        WHERE related_publication_id = "'.$id.'"
+        AND name = "'.$name.'"');
 
-    $q2 = $this->_db->query('SELECT user_id FROM publication WHERE id = "'.$id.'"');
-    $donnees = $q2->fetch(PDO::FETCH_ASSOC);
-    $user_manager = new UsersManager($this->_db);
-    $user = $donnees->get($donnees['user_id']);
-    
-    $q3 = $this->_db->prepare('UPDATE stat SET value = value-1
-      WHERE related_user_id = "'.$user.'"
-      AND name = "'.$name.'"');
+      $q2 = $this->_db->query('SELECT user_id FROM publication WHERE id = "'.$id.'"');
+      $donnees = $q2->fetch(PDO::FETCH_ASSOC);
+      $user_manager = new UsersManager($this->_db);
+      $user = $user_manager->get($donnees['user_id']);
+      
+      $q3 = $this->_db->prepare('UPDATE stat SET value = value-1
+        WHERE related_user_id = "'.$user->get_id().'"
+        AND name = "'.$name.'"');
 
-    // change la table vote
-    $q4 = $this->_db->prepare('DELETE FROM vote
-      WHERE user_id = "'.$user.'"
-      AND publication_id = "'.$id.'"
-      AND name = "'.$name.'"');
-    
-    DELETE FROM vote WHERE vote.id = 1
+      // change la table vote
+      $q4 = $this->_db->prepare('DELETE FROM vote
+        WHERE user_id = "'.$user->get_id().'"
+        AND publication_id = "'.$id.'"
+        AND name = "'.$name.'"');
 
-    $q1->execute();
-    $q2->execute();
-    $q3->execute();
+      $q1->execute();
+      $q3->execute();
+      $q4->execute();
+      echo "ok le vote a été supprimé";
+    }
+    else {
+      echo "vous n'avez pas voté et ne pouvez pas annuler votre vote";
+      return false; // l'user n'a pas voté sur cette stat
+    }
   }
 
   public function update(Stat $stat)
